requirements.txt
httpx[http2]>=0.27,<1
transnetv2-pytorch>=1.0.5,<2
transformers>=4.57,<6
accelerate>=1.1,<2
torch>=2.0,<3
torchvision>=0.20,<1
torchcodec>=0.1,<1
pandas>=2.2,<3
pillow>=11,<12
timm>=1.0.11,<2
wdtagger>=0.16,<1
